+++ src/main/scala/edu/ie3/simona/model/participant2/FixedFeedInModel.scala
@@ -26,6 +26,7 @@ import edu.ie3.simona.model.participant2.ParticipantModel.{
 import edu.ie3.simona.service.ServiceType
 import edu.ie3.util.quantities.PowerSystemUnits
 import edu.ie3.util.quantities.QuantityUtils.RichQuantityDouble
+import edu.ie3.util.scala.quantities.QuantityConversionUtils.PowerConversionSimona
 import edu.ie3.util.scala.quantities.{ApparentPower, Kilovoltamperes}
 
 import java.time.ZonedDateTime
@@ -102,12 +103,7 @@ object FixedFeedInModel {
       new FixedFeedInModel(
         input.getUuid,
         input.getId,
-        Kilovoltamperes(
-          input.getsRated
-            .to(PowerSystemUnits.KILOVOLTAMPERE)
-            .getValue
-            .doubleValue
-        ),
+        input.getsRated.toApparent,
         input.getCosPhiRated,
         QControl.apply(input.getqCharacteristics),
       )