+++ src/main/scala/edu/ie3/simona/model/participant2/load/FixedLoadModel.scala
@@ -9,21 +9,20 @@ package edu.ie3.simona.model.participant2.load
 import edu.ie3.datamodel.models.input.system.LoadInput
 import edu.ie3.simona.config.RuntimeConfig.LoadRuntimeConfig
 import edu.ie3.simona.model.participant.control.QControl
-import edu.ie3.simona.model.participant2.ParticipantModel
 import edu.ie3.simona.model.participant2.ParticipantModel.{
   ActivePowerOperatingPoint,
   FixedState,
-  OperatingPoint,
   ParticipantFixedState,
   ParticipantModelFactory,
 }
 import edu.ie3.simona.service.ServiceType
-import edu.ie3.util.quantities.PowerSystemUnits
-import edu.ie3.util.quantities.PowerSystemUnits.KILOWATTHOUR
-import edu.ie3.util.scala.quantities.{ApparentPower, Kilovoltamperes}
-import squants.time.Days
+import edu.ie3.util.scala.quantities.ApparentPower
+import edu.ie3.util.scala.quantities.QuantityConversionUtils.{
+  EnergyToSimona,
+  PowerConversionSimona,
+}
 import squants.Power
-import squants.energy.KilowattHours
+import squants.time.Days
 
 import java.time.ZonedDateTime
 import java.util.UUID
@@ -63,20 +62,13 @@ object FixedLoadModel {
     override def create(): FixedLoadModel = {
       val referenceType = LoadReferenceType(config.reference)
 
-      val sRated = Kilovoltamperes(
-        input.getsRated
-          .to(PowerSystemUnits.KILOVOLTAMPERE)
-          .getValue
-          .doubleValue
-      )
+      val sRated = input.getsRated.toApparent
 
       val activePower: Power = referenceType match {
         case LoadReferenceType.ACTIVE_POWER =>
           sRated.toActivePower(input.getCosPhiRated)
         case LoadReferenceType.ENERGY_CONSUMPTION =>
-          val eConsAnnual = KilowattHours(
-            input.geteConsAnnual().to(KILOWATTHOUR).getValue.doubleValue
-          )
+          val eConsAnnual = input.geteConsAnnual().toSquants
           eConsAnnual / Days(365d)
       }