+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalHouse.scala
@@ -88,7 +88,7 @@ final case class ThermalHouse(
     * change in external feed in will take place.
     *
     * @param currentThermalHouseState
-    *   Most recent state, that is valid for this model.
+    *   Actual state, that is valid for this model.
     * @return
     *   The needed energy in the questioned tick.
     */
@@ -381,7 +381,7 @@ object ThermalHouse {
 
   def startingState(house: ThermalHouse): ThermalHouseState =
     ThermalHouseState(
-      -1L,
+      0L,
       house.targetTemperature,
       zeroMW,
     )